Move symbol tests back into tests/ for now
